Make target model non-trainable + Scoping + l2_loss
This is a rather big change.

Makes all of the target model parameters non-trainable. It is unclear if
they originally had an effect on the gradient calculation, but now they
don't for sure.

Adds variable scoping, making for a much clearer graph in tensorboard.

Changes the loss from self concocted mean + pow to tf.nn.l2_loss